// src/app/models/paciente.models.ts
export interface Paciente {
    key: string;
    _id: number;
    comentario: string;
    nombre: string;
    apellidos: string;
    tipo_documento: string;
    documento_identidad: string;
    fecha_nacimiento: string;
    sexo: string;
    calle: string;
    telefono: string;
    telefono_adicional: string;
    email: string;
    numero: string;
    puerta: string;
    pais: string;
    ciudad: string;
    cp: string;
    aseguradora: string;
    numero_seguro: string;
    otros_datos: string;
    fecha_registro: string;
    photo: string;
    LPD: string;
    alergias: string;
}